core: reorg filetree
