+++ core/src/main/kotlin/ch/kleis/lcaac/core/lang/register/DataRegister.kt
@@ -1,0 +1,10 @@
+package ch.kleis.lcaac.core.lang.register
+
+import ch.kleis.lcaac.core.lang.expression.DataExpression
+
+data class DataKey(
+    val name: String,
+) {
+    override fun toString() = name
+}
+typealias DataRegister<Q> = Register<DataKey, DataExpression<Q>>
