+++ core/src/main/kotlin/ch/kleis/lcaac/core/lang/register/DimensionRegister.kt
@@ -1,0 +1,10 @@
+package ch.kleis.lcaac.core.lang.register
+
+import ch.kleis.lcaac.core.lang.dimension.Dimension
+
+data class DimensionKey(
+    val name: String,
+){
+    override fun toString() = name
+}
+typealias DimensionRegister = Register<DimensionKey, Dimension>
