+++ core/src/main/kotlin/ch/kleis/lcaac/core/lang/register/Register.kt
@@ -1,11 +1,6 @@
-package ch.kleis.lcaac.core.lang
+package ch.kleis.lcaac.core.lang.register
 
 import arrow.optics.Fold
-import ch.kleis.lcaac.core.lang.dimension.Dimension
-import ch.kleis.lcaac.core.lang.expression.DataExpression
-import ch.kleis.lcaac.core.lang.expression.EProcessTemplate
-import ch.kleis.lcaac.core.lang.expression.ESubstanceCharacterization
-import ch.kleis.lcaac.core.lang.expression.SubstanceType
 
 class Register<K, E> (
     private val data: Map<K, E> = HashMap()
@@ -66,44 +61,3 @@ class Register<K, E> (
     }
 }
 
-data class DataKey(
-    val name: String,
-) {
-    override fun toString() = name
-}
-typealias DataRegister<Q> = Register<DataKey, DataExpression<Q>>
-
-data class DimensionKey(
-    val name: String,
-){
-    override fun toString() = name
-}
-typealias DimensionRegister = Register<DimensionKey, Dimension>
-
-data class ProcessKey(
-    val name: String,
-    val labels: Map<String, String> = emptyMap(),
-){
-    override fun toString(): String {
-        return name + labels.entries.joinToString(", ") { "${it.key}=${it.value}" }
-            .let { if (it.isNotEmpty()) "{$it}" else "" }
-    }
-}
-typealias ProcessTemplateRegister<Q> = Register<ProcessKey, EProcessTemplate<Q>>
-
-data class SubstanceKey(
-    val name: String,
-    val type: SubstanceType? = null,
-    val compartment: String? = null,
-    val subCompartment: String? = null,
-) {
-    override fun toString(): String {
-        return name + listOfNotNull(
-            type?.let { "type=$it" },
-            compartment?.let { "compartment=$it" },
-            subCompartment?.let { "sub_compartment=$it" },
-        ).joinToString(", ").let { if (it.isNotEmpty()) "($it)"  else ""}
-    }
-}
-
-typealias SubstanceCharacterizationRegister<Q> = Register<SubstanceKey, ESubstanceCharacterization<Q>>
