+++ core/src/test/kotlin/ch/kleis/lcaac/core/lang/evaluator/EvaluatorTest.kt
@@ -4,6 +4,10 @@ import ch.kleis.lcaac.core.lang.*
 import ch.kleis.lcaac.core.lang.dimension.UnitSymbol
 import ch.kleis.lcaac.core.lang.expression.*
 import ch.kleis.lcaac.core.lang.fixture.*
+import ch.kleis.lcaac.core.lang.register.ProcessKey
+import ch.kleis.lcaac.core.lang.register.ProcessTemplateRegister
+import ch.kleis.lcaac.core.lang.register.SubstanceCharacterizationRegister
+import ch.kleis.lcaac.core.lang.register.SubstanceKey
 import ch.kleis.lcaac.core.lang.value.*
 import ch.kleis.lcaac.core.math.basic.BasicNumber
 import ch.kleis.lcaac.core.math.basic.BasicOperations
