+++ plugin/src/main/kotlin/ch/kleis/lcaac/plugin/language/psi/type/PsiProcess.kt
@@ -1,10 +1,9 @@
 package ch.kleis.lcaac.plugin.language.psi.type
 
-import ch.kleis.lcaac.core.lang.ProcessKey
+import ch.kleis.lcaac.core.lang.register.ProcessKey
 import ch.kleis.lcaac.plugin.language.psi.stub.process.ProcessStub
 import ch.kleis.lcaac.plugin.language.psi.type.trait.BlockMetaOwner
 import ch.kleis.lcaac.plugin.psi.*
-import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNameIdentifierOwner
 import com.intellij.psi.StubBasedPsiElement
